Update to platform/SDK 28 + version bump deps.
This might not update everything, but it does get to
a working configuration on SDK/platform v28.

This will bump the minSdkVersion from 15 to 16. According to
https://developer.android.com/about/dashboards/index.html
this will make CarStream no longer work on Android 4.0.4 and
earlier devices. 4.1.x and later should still work. This
represents about 0.6% of the active devices in the last
7 days (0.3% for the 4.0 family plus 0.3% for the 2.3 family).
One of those is probably mine, and I can't imagine actually
wanting to use the thing to connect to Android Auto; maps
would be sooooo slow, assuming any of it would work at all.