Build/crash fix: Didn't find class "okio.Buffer"
okio is a dependency of okhttp3. Presumably, it wasn't necessary
for this dependency to be specified in the earlier Android SDK
version (not needed on 26, maybe needed on 28), or perhaps this
was pulled by a newer version of okhttp3 (not needed for
okhttp3-3.9.1, maybed needed for okhttp3-4.0.0-alpha01). At any
rate, this commit prepares the codebase for such version bumps.

I believe I encountered this as a runtime crash, though this
problem, by its nature, could be encountered at build time
depending on configuration:
https://talk.remobjects.com/t/build-error-when-including-com-squareup-okio-jar/18249

The error message encountered was something like this:

  java.lang.ClassNotFoundException: Didn't find class "okio.Buffer"

(It was longer, but I can't be more specific because I no longer
have the log for it.)