Build fix: "send_to_car" not in default locale
When updating to SDK version 28, an error like this would appear
in the build log:
~~~~~~
C:\Users\z\StudioProjects\carstream-android-auto\app\src\main\res\values-ro\strings.xml:14: Error: "send_to_car" is translated here but not found in default locale [ExtraTranslation]
    <string name="send_to_car">Trimitere la masina</string>
            ~~~~~~~~~~~~~~~~~~

   Explanation for issues of type "ExtraTranslation":
   If a string appears in a specific language translation file, but there is
   no corresponding string in the default locale, then this string is probably
   unused. (It's technically possible that your application is only intended
   to run in a specific locale, but it's still a good idea to provide a
   fallback.).

   Note that these strings can lead to crashes if the string is looked up on
   any locale not providing a translation, so it's important to clean them
   up.
~~~~~~

This commit resolves the issue by adding a send_to_car entry
in the values/strings.xml file with the value "Send to car".
This might be similar in nature to the "receive_from_phone",
just different in that the string isn't explicitly referenced
in the project. I don't like the idea of potential crashes,
so I'm putting a defualt value in English and hoping that its
identifier is representative of what it's supposed to expand
to.